#Navigating_a_ResultSet
#Navigating_a_ResultSet

There are several methods in the ResultSet interface that involve moving the cursor, including −
S.N.--------->Methods & Description

1 	public void beforeFirst() throws SQLException 
Moves the cursor just before the first row.

2 	public void afterLast() throws SQLException
Moves the cursor just after the last row.

3 	public boolean first() throws SQLException
Moves the cursor to the first row.

4 	public void last() throws SQLException
Moves the cursor to the last row.

5 	public boolean absolute(int row) throws SQLException
Moves the cursor to the specified row.

6 	public boolean relative(int row) throws SQLException
Moves the cursor the given number of rows forward or backward, from where it is currently pointing.

7 	public boolean previous() throws SQLException
Moves the cursor to the previous row. This method returns false if the previous row is off the result set.

8 	public boolean next() throws SQLException
Moves the cursor to the next row. This method returns false if there are no more rows in the result set.

9 	public int getRow() throws SQLException
Returns the row number that the cursor is pointing to.

10 	public void moveToInsertRow() throws SQLException
Moves the cursor to a special row in the result set that can be used to insert a new row into the database. The current cursor location is remembered.

11 	public void moveToCurrentRow() throws SQLException
Moves the cursor back to the current row if the cursor is currently at the insert row; otherwise, this method does nothing